move test data to separate directory
